splitted up decoders to core and ds
