+++ geo_skeletons/decoders/__init__.py
@@ -1,2 +1,3 @@
-from .decoders import identify_core_in_ds, map_ds_to_gp, core_dicts_from_ds
+from .core_decoders import identify_core_in_ds, core_dicts_from_ds
+from .ds_decoders import map_ds_to_gp
 from .setters import set_core_vars_to_skeleton_from_ds, add_dynamic_vars_from_ds
